Create 07
You will receive an unknown number of lines. On each line, you will receive an array with 3 elements. The first
element will be string and represents the name of the person. The second element will be a string and will
represent the ID of the person. The last element will be an integer which represents the age of the person.
When you receive the command &quot;End&quot;, stop taking input and print all the people, ordered by age.